Move ABI to src directory
