+++ package.json
@@ -56,7 +56,8 @@
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "eject": "react-scripts eject",
+    "typechain": "typechain --target=ethers-v5 --outDir=./src/contracts/ ./src/abi/*.json"
   },
   "eslintConfig": {
     "extends": [
@@ -77,9 +78,11 @@
     ]
   },
   "devDependencies": {
+    "@typechain/ethers-v5": "^6.0.5",
     "@types/react-notifications-component": "^3.1.0",
     "@types/react-router-dom": "^5.1.7",
     "@types/react-table": "^7.0.29",
-    "@types/styled-components": "^5.1.9"
+    "@types/styled-components": "^5.1.9",
+    "typechain": "^4.0.3"
   }
 }
